Comenzar a usar AdminLTE con ayuda de Libman y NPM Darle una actualización a UI Ajustes visuales a todas las pantallas Preparar el terreno para la llegada del dashboard
