+++ src/Cashflowio.Web/Startup.cs
@@ -89,7 +89,7 @@ namespace Cashflowio.Web
             {
                 routes.MapRoute(
                     name: "default",
-                    template: "{controller=Dashboard}/{action=Transfer}/{id?}");
+                    template: "{controller=Dashboard}/{action=Index}/{id?}");
             });
         }
     }
