+++ src/Cashflowio.Web/Views/Dashboard/Index.cshtml
@@ -1,0 +1,3 @@
+﻿@{
+    ViewBag.Title = "Dashboard";
+}
