+++ src/Cashflowio.Web/Views/IncomeSource/Index.cshtml
@@ -19,7 +19,7 @@
 }
 
 <ejs-grid id="Grid" toolbar="@(new List<string> {"Add", "Edit", "Delete"})"
-          height="555" width="auto" allowTextWrap="true" allowPaging="true" allowSorting="true">
+          height="420" width="auto" allowTextWrap="true" allowPaging="true" allowSorting="true">
     <e-grid-pagesettings pageSize="50" pageSizes="@(new[] {"10", "50", "100", "200", "500", "1000", "All"})"></e-grid-pagesettings>
     <e-grid-filtersettings type="Excel"></e-grid-filtersettings>
     <e-grid-sortsettings columns="@(new[] {new {field = "name", direction = "Ascending"}})"></e-grid-sortsettings>
