+++ src/Cashflowio.Web/Views/RawTransaction/Grid.cshtml
@@ -4,7 +4,7 @@
 }
 
 <ejs-grid id="Grid" dataSource="@Model"
-          enableHover="false" height="555" width="auto" allowSelection="false"
+          enableHover="false" height="420" width="auto" allowSelection="false"
           toolbar="@(new List<string> {"Search"})" allowTextWrap="true"
           allowPaging="true" allowSorting="true" allowFiltering="true" load="onload">
     <e-grid-pagesettings pageSize="50" pageSizes="@(new[] {"50", "100", "200", "500", "1000", "All"})"></e-grid-pagesettings>
